requirements.txt
fastapi>=0.110.0,<1.0.0
pydantic-settings>=2.0.0,<3.0.0
uvicorn[standard]>=0.27.0,<1.0.0
weaviate-client>=4.9.6,<5.0.0
sentence-transformers>=3.0.1,<4.0.0

